Update 04-end-of.md
Confused if it was a 'date' string or a hard-coded 'string'.  

An additional note that the suggested change doesn't take into account because I wasn't sure how it would work, is if the unit of time is minutes - what would it do? Would it be 59 or 60? If it is restricted to certain units, should they be listed?